+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Release Notes for Campaign
 
+## 1.23.6 - 2022-11-07
+### Changed
+- Test requests from Mailgun now return a success response.
+
+### Fixed
+- Fixed a bug in which the Mailgun webhook controller action was not processing requests correctly ([#341](https://github.com/putyourlightson/craft-campaign/issues/341)).
+
 ## 1.23.5 - 2022-10-12
 ### Fixed
 - Fixed a bug in which blocked contacts were being considered as pending contacts, meaning that sendouts could hang during sending ([#324](https://github.com/putyourlightson/craft-campaign/issues/324)).